LayoutTests/imported/w3c/web-platform-tests/html/dom/documents/resource-metadata-management/document-lastModified-expected.txt


FAIL lastModified should return the last modified date and time assert_equals: expected "12/31/1969 17:23:45" but got "01/01/1970 01:23:45"
